Refactor: Move dto package content to model package
Moved Member.kt, Team.kt, and ModelMappers.kt from org.ole.planet.myplanet.model.dto to org.ole.planet.myplanet.model.
Updated package declarations and imports throughout the codebase.
Removed the now empty dto directory.
