+++ app/src/main/java/org/ole/planet/myplanet/base/BaseTeamFragment.kt
@@ -13,7 +13,7 @@ import kotlinx.coroutines.withContext
 import org.ole.planet.myplanet.model.RealmMyTeam
 import org.ole.planet.myplanet.model.RealmNews
 import org.ole.planet.myplanet.model.RealmUser
-import org.ole.planet.myplanet.model.dto.Team
+import org.ole.planet.myplanet.model.Team
 import org.ole.planet.myplanet.repository.TeamsRepository
 
 @AndroidEntryPoint
