+++ app/src/main/java/org/ole/planet/myplanet/model/ChatShareTargets.kt
@@ -1,7 +1,5 @@
 package org.ole.planet.myplanet.model
 
-import org.ole.planet.myplanet.model.dto.Team
-
 data class ChatShareTargets(
     val community: Team?,
     val teams: List<RealmMyTeam>,
